Move from UI5 sources to generic sources
We currently aren't enabling standard queries to use UI5 remote flow
sources.
This is changed in this commit where we move the UI5 sources into the
generic remote flow source concept to enable all queries to be able to
use them.

This also means we can no longer differentiate between between the
origin of sources,
so our UI5 specific queries become more general in that any remote flow
source to a UI5 specific sink is reported.

This changes the definition of a UI5 specific query such that it is the
sink that differentiate it from our standard library query so we can
keep providing UI5 specific remediations in the documentation.