Another attempted fix for index-files.js cwd
Changes the way the index-files.js script is invoked such that the
original `--source-root` directory is used, where possible, as the
current working directory for any work performed with within the
extractor. Passes the original working directory as a parameter of
the index-files.js script to allow that (child) script to run from
the project/source root while ensuring node package dependencies
are still installed in `extractors/cds/tools/node_modules`.