index-files.js must compiles cds to file (not dir)
Forces the `cds` compilier to output JSON to a file (via stdout) instead
of creating an output directory. Accounts for what appears to be a
change in the behavior of the `cds` (CLI) compiler, where the `-o` (or
`--dest`) option now `Writes output to the given folder instead of stdout`.

```
$ npx cds --version

@cap-js/asyncapi: 1.0.2
@cap-js/db-service: 1.16.0
@cap-js/openapi: 1.1.2
@cap-js/sqlite: 1.7.7
@capire/bookshop: 1.0.0
@sap/cds: 8.5.0
@sap/cds-compiler: 5.2.0
@sap/cds-dk: 8.7.1
@sap/cds-fiori: 1.2.7
@sap/cds-foss: 5.0.1
@sap/cds-mtxs: 2.5.1
@sap/eslint-plugin-cds: 3.1.2
Node.js: v20.15.0
```