Remove UI5AMDModule
This one caused a new non-monotonic recursion; it
is not necessary at the moment since it is not
used anywhere apart from a vacuous `import`.
The AMD-style module system is already defined in
UI5.qll as abstract class `UserModule`; so deleting
this won't break the rest of the UI5 queries.