Make `SapDefineModule` extend `AmdModuleDefinition::Range` and update the expected results
`SapDefineModule` in its previous form did not extend
`AmdModuleDefinition::Range` which in turn extends standard library's
`Module` that enables its subclasses to be identified with MaD definitions.
The reason the queries have been working was thanks to `UI5AMDModule.qll` that
provided a `SapAmdModule` class extending the `Module` class directly.

The problem with `SapAmdModule` copy-pasted and only slightly modified
the `AMD.qll` in the standard library, so it was out of sync with the
standard library couterpart when the DataFlow library behind it was
overhauled to have a new API. It was causing the majority of the
problems when the `qlpack.yml`s were updated with the latest DataFlow
API and the `SapAmdModule` failed to play nicely with the updated
library modules, emitting new non-monotonic recursion errors.

Therefore, this commit makes the `SapDefineModule` in `UI5.qll` extend
`AmdModuleDefinition::Range` and removes the outdated `SapAmdModule`.
`SapDefineModule`s are AMD-style modules defined with
`sap.ui.define` or `sap.ui.require` function calls that augments the
global AMD `define` functions with the capability of extending another
such module.