Run 'cds compile' command without '-o'
Updates unit-test and cds-extractor use of the `cds compile` command in
order to avoid the use of the `-o` option, which outputs to a directory
instead of a file, as intended. Replaces use of the `-o` option with a
simple redirect of cds compiler output to stdout, which is redirected to
the indended '.cds.json' file path.

This commit should resolve failures for both code scanning and unit test
workflows for this project/repo.