Avoid redirecting cds compiler stdout
Revokes the approach of redirecting stdout for 'cds compile' commands
used in the CDS extractor. In some cases, stdout may contain more than
just the contents of the generated JSON, which creates errors that are
difficult to troubleshoot as they only occur in some environments.

Returns to the approach of using CLI arguments to instruct the CDS
compiler to output to a specified path, and adds some logic to account
for variations in CDS compiler behavior when using the '-o' or '--dest'
CLI argument. Because some CDS compiler versions say that they output
to a directory but actually output to a file, and some CDS compiler
versions actually output to a directory as specified in the CLI docs,
we trust the use of the '--dest' argument (more than we trust a redirect
of stdout to a JSON file), but we *verify* whether the output is a
regular '.cds.json' file or is, indeed, a directory.