Replace old index-files.js with .ts equivalent
Replaces the JS-based indes-files script with a TS-based index-files
script, and also sets up TypeScript (TS) build command with eslint
support such that the compiling the `index-files.ts` script will
create (one giant) `index-files.js` script (with a `index-files.js.map`
companion). Thus, sets up a TS-based equivalent of the index-files.js
CDS extractor script as preparation for a broader rewrite of the CDS
extractor to be more performant and maintainable.

Adds a documentation file, `autobuild.md`, to help document and guide
the multi-stage process of rewriting the CDS extractor to use an
autobuild-based approach, where this commit represents the first stage
of the multi-stage process.

Adds `.gitignore` files in subdirectories where testing the CDS
extractor tends to create changes that we never want to commit.