Refactor index-files.ts script for modularity
Refactors the index-files.ts script to be follow an object-orient
programming design, with the bulk of the script logic abstracted into
re-useable functions in the new `extractors/cds/tools/src` directory.

Adds a `.prettierrc.js` and updates the `.eslintrc.js` and
`package.json` files for the CDS extractor in order to provide better
linting coverage / rules for the TypeScript (`.ts`) code for the CDS
extractor sub-project.