Fix2 for issues identified in peer review
Improves error handling during initial environment setup for the CDS
extractor. Ensures that error codes are used when exiting due to an
error in the configuration of the CDS extractor.

Minor refactoring of index-files.ts script in order to migrate
environment setup tasks to dedicated functions with associated unit
tests.

Migrates response-file checking and related logic to a dedicated
'getCdsFilePathsToProcess` function, which also helps to prepare for
dropping the "index-files" approach in favor of an "autobuild" based
approach.