Rename CDS extractor entrypoint and refactor args
Renames the entrypoint to the CDS extractor script and refactors its
arguments in order to support using different "run modes" for the
extractor, including:

  - "autobuild" : work-in-progress, just a stub right now;
  - "debug-parser" : using for debugging CDS project & file parsing;
  - "index-files" : legacy mode, useful for backwards compatibility;

Updates the usage (help) message for the script to represent the
required arguments for each of the currently planned run modes.

Adds support for the "debug-parser" run mode, which debugs to a file
under the `extractors/cds/tools/out/debug/` directory. Useful for
in-progress rewrite of the CDS extractor to be more performant when
running and more useful in terms of yielding a CodeQL database that
allows for high-precision query results for CDS projects/queries.