Simplify CDS extractor logic and refactor
Refactors cds extractor `src/cds/compiler` and `src/cds/parser`
packages for improved maintainability.

Simplifies the main logic of the CDS extractor such that we always
build a graph that maps CDS projects to their imports / dependencies,
which is part of the longer process of deprecating the "index-files"
run mode of the CDS extractor (in favor of autobuild, eventually).

Attempts to fix CDS file and project parsing for test projects such as:
`javascript/frameworks/cap/test/queries/loginjection/log-injection-without-protocol-none`