Fix project-aware CDS compile file paths
Fixes a regression where the project base directory was being used
to set the `cwd` of the process spawned for running the CDS compiler
for "project-aware" compilation. Adds unit tests to ensure the `cwd`
is always set to the value of the `sourceRoot` directory.

Further refactoring of the `cds/compiler` and `cds/parser` packages
within the source code of the CDS extractor.

This commit is expected to actually cause more problems with existing
queries, despite fixing the relative-file-path problem / regression.
Some changes to existing CodeQL queries and/or expected results may be
required as, at this point, the JSON data generated by the CDS compiler
(via the CDS extractor) seems valid.