Replace Math.random() in CDS extractor
Addresses a code scanning alert related to the recent introduction
of the `Math.random()` function when creating a session ID for CDS
extractor logging. Replaces `Math.random()` with `Data.now()` in an
effort to remove any perception that the "session ID" is in any way
used in a "security context".