Add documentation, make `CdsDb` extend `CdsDbService`, and remove `isFromCdsLib/0`
As stated in the new docstring of `CdsFacade`, the
same `cds_facade` object can be accessed either via
`@sap/cds` or `@sap/cds/lib`. Therefore, there is
no distinction to be made between the two.

Also, the `CdsDb` and the `GloballyAccessedCdsDbService`
aimed to capture the same thing, so delete the latter
in favor of the former (it's clearer and more concise).