Setup CDS extractor esbuild JS bundle
Changes the build process for the CDS extractor to produce
`dist/cds-extractor.bundle.js` and `dist/cds-extractor.bundle.js.map`
files instead of a pair of `.js` and `.js.map` files for every
compiled `.ts` file.

Allows for pre-build of the JavaScript code used to run the CDS
extractor while committing the bare minimum number of files for the
"distribution" of the build.