Use 'node20' as target for CDS extractor build
Standardizes the CDS extractor build configs to use a 'node20' as
the JS runtime "target".

Updates the CDS extractor's package.json file in order to specify
required "engines" for `node` and `npm`.

Minimizes the `tsconfig.json` config used by the CDS extractor in
order to only specify configs required for linting and testing,
given that the main build config is now specified in the
`esbuild.config.mjs` file.