Implement retry for CDS compilation tasks
This commit updates the implementation of the CDS extractor
in order to attempt to retry failed compilation tasks using
the full set of node dependencies specified by the
`package.json` config file of the associated CDS project.

The new implementation continues to use a minimal cache of
`@sap/cds` and `@sap/cds-dk` dependencies as the default/initial
installation behavior as a performance optimization. When this
does not work, the CDS extractor should now attempt to install
all node dependencies for the associated (CAP) project before
re-attempting the failed compilation tasks(s) for that project.

This commit also updates the `package.json` config for the CDS
extractor itself in order to ensure that we build test code
before running unit tests, which addresses a bug in the previous
testing implementation. The test code is now automatically
compiled as part of running commands like `npm run test` or
`npm run test:coverage`.