Rework CDS project-only compilation & CDL.qll
Reworks the CDS extractor such that it only does project-level
compilation, which allows for consistently generating a single
model.cds.json file in the base/root directory of its associated
project, which allows for matching file paths relative to the
project base directory.

Updates the `CDL.qll` library file to use location/path values
that are relative to the project base directory rather than relative
to the source root directory.