Enforce min version for @sap/cds-dk dependency
For CDS compilation tasks that need to be retried because the
initial attempt failed to produce a `model.cds.json` file, this
commit enforces that the `npx cds compile` command will require
a compatible, non-deprecated version of the `@sap/cds-dk`
dependency by using the `--yes --package @sap/cds-dk@^8` arguments
with the `npx cds compile` command.

This addresses a persistent challenge with reliably performing
CDS compilation tasks via some form of the `npx cds` command when
run from the base directory of a given project. If the `npx` command
does not enforce the use of a compatible `--package @sap/cds-dk@*`,
then we sometimes encounter errors in environments where the project
inherits/uses a non-compatible version of `@sap/cds-dk` from the
`node_modules` of some parent directory of the project.